+++ src/components/edit/thematic/ThematicDialog.js
@@ -6,6 +6,7 @@ import { SystemSettingsCtx } from '../../SystemSettingsProvider';
 import { Tab, Tabs } from '../../core';
 import ValueTypeSelect from './ValueTypeSelect';
 import AggregationTypeSelect from './AggregationTypeSelect';
+import CompletedOnlyCheckbox from './CompletedOnlyCheckbox';
 import NoDataColor from './NoDataColor';
 import DataElementGroupSelect from '../../dataElement/DataElementGroupSelect';
 import DataElementSelect from '../../dataElement/DataElementSelect';
@@ -422,6 +423,7 @@ export class ThematicDialog extends Component {
                                 ),
                             ]}
                             <AggregationTypeSelect className={styles.select} />
+                            <CompletedOnlyCheckbox valueType={valueType} />
                         </div>
                     )}
                     {tab === 'period' && (